+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+## 29/05/2020
+
+- Add endpoint for testing subscription email alerts for GLAD alerts and VIIRS Fires. [#172772548](https://www.pivotaltracker.com/story/show/172772548)
+
 # v.1.4.2
 
 ## 19/05/2020